Enable logging

+++ chatai/sql/tables.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import BigInteger, Column, Integer, String
 from sqlalchemy.ext.declarative import declarative_base
 
 
@@ -7,10 +7,10 @@ Base = declarative_base()
 class Message(Base):
     __tablename__ = 'messages'
 
-    id = Column(Integer, primary_key=True, autoincrement=False)
-    chat_id = Column(Integer, nullable=True)
+    id = Column(BigInteger, primary_key=True, autoincrement=False)
+    chat_id = Column(BigInteger, nullable=True)
     username = Column(String, nullable=True)
     text = Column(String, nullable=True)
     unixtime = Column(Integer, nullable=True)
     image_b64_encoded = Column(String, nullable=True)
-    reply_to_message_id = Column(Integer, nullable=True)
+    reply_to_message_id = Column(BigInteger, nullable=True)
